lesson13

+++ src/test/groovy/pl/futurecollars/invoicing/controller/AbstractControllerTest.groovy
@@ -63,11 +63,11 @@ class AbstractControllerTest extends Specification {
                 invoice
             }
         }
-        void deleteInvoice(int id) {
+        void deleteInvoice(long id) {
             mockMvc.perform(delete("$INVOICE_ENDPOINT/$id"))
                     .andExpect(status().isNoContent())
         }
-        Invoice getInvoiceById(int id) {
+        Invoice getInvoiceById(long id) {
             def invoiceAsString = mockMvc.perform(get("$INVOICE_ENDPOINT/$id"))
                     .andExpect(status().isOk())
                     .andReturn()
@@ -75,7 +75,7 @@ class AbstractControllerTest extends Specification {
                     .contentAsString
             jsonService.toObject(invoiceAsString, Invoice)
         }
-        String invoiceAsJson(int id) {
+        String invoiceAsJson(long id) {
             jsonService.toJson(invoice(id))
         }
